# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Api to manage clients favorites products",
        "title": "Favorites Api",
        "contact": {
            "name": "API Support",
            "url": "https://www.linkedin.com/in/rspandre/",
            "email": "rsp.assistencia@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/v1",
    "paths": {
        "/admin/clients": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit to list clients",
                "tags": [
                    "Admin"
                ],
                "summary": "List Client",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ClientResumeResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit to register client",
                "tags": [
                    "Admin"
                ],
                "summary": "Client register",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClientRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/admin/clients/{clientID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit client detail",
                "tags": [
                    "Admin"
                ],
                "summary": "Client Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Client id",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClientDetailResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit client update",
                "tags": [
                    "Admin"
                ],
                "summary": "Client Update",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Client id",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClientRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/clients": {
            "post": {
                "description": "Endpoit to self register",
                "tags": [
                    "Client"
                ],
                "summary": "Client register",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClientRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/clients/{clientID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit client detail",
                "tags": [
                    "Client"
                ],
                "summary": "Client Detail",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Client id",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ClientDetailResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit client update",
                "tags": [
                    "Client"
                ],
                "summary": "Client Update",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Client id",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.ClientRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit client delete",
                "tags": [
                    "Admin"
                ],
                "summary": "Client Delete",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 1,
                        "description": "Client id",
                        "name": "clientID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/favorites": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit add favorite",
                "tags": [
                    "Favorite"
                ],
                "summary": "Add Favorite",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.FavoriteRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/favorites/{clientId}/{productId}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit delete favorite",
                "tags": [
                    "Favorite"
                ],
                "summary": "Delete Favorite",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "client id",
                        "name": "clientId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "productId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "Endpoit login",
                "tags": [
                    "Login"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Payload",
                        "name": "Payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LoginRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.LoginResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/products": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit to list products",
                "tags": [
                    "Product"
                ],
                "summary": "list products",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page number",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.ProductResponseDTO"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        },
        "/products/{productID}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Endpoit to detail product",
                "tags": [
                    "Product"
                ],
                "summary": "detail product",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "product id",
                        "name": "productID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.ProductResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/genericresponse.Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.ClientDetailResponseDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "favorites": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ProductResponseDTO"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ClientRequestDTO": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.ClientResumeResponseDTO": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "dto.FavoriteRequestDTO": {
            "type": "object",
            "required": [
                "clientId",
                "productId"
            ],
            "properties": {
                "clientId": {
                    "type": "integer"
                },
                "productId": {
                    "type": "integer"
                }
            }
        },
        "dto.LoginRequestDTO": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "dto.LoginResponseDTO": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "dto.ProductResponseDTO": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                },
                "reviewScore": {
                    "type": "number"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "genericresponse.Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}